backend/bot/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio
sqlite3
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2

